feat: base implementation of client and server with node registration
Provides the PoC baseline for Supernetes: the HPC client is able to connect to the server running in the cloud environment via gRPC, and send a (dummy) list of HPC nodes to the server. The server then registers these nodes in the Kubernetes cluster by starting Virtual Kubelet node controllers. As the node availability changes, the server automatically reconciles the client's updates. TODO: node deletion is not quite working yet, stale nodes seem to persist in the cluster.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>