ci: configure environment for deployment
Deployment to the container registry from PRs from forks can't be done with`pull_request` due to lack of secret access. We can work around this by triggering `workflow_run`, which runs with the target repo's privileges. To avoid a security risk, introduce an additional dummy "Approve" workflow, which requires explicitly approved access to an environment.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>