ci: switch to fork of rootless Docker Action
Upgrading to `ubuntu-24.04` is blocked by this Action due to [AppArmor changes](https://github.com/ScribeMD/rootless-docker/issues/401). The [PR for fixing this](https://github.com/ScribeMD/rootless-docker/pull/402) is up, but upstream is dormant. However, Renovate will try the `ubuntu-24.04` upgrade immediately, so let's switch to a pinned version of the Action that carries the fix for now.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>